Moved HadoopRecordReader into HadoopInputFormat
