+++ .circleci/config.yml
@@ -71,12 +71,14 @@ jobs:
       - install-bazel-linux-rbe
       - checkout
       - run: |
-          SONARCLOUD_CODE_ANALYSIS_CREDENTIAL=$SONARCLOUD_CREDENTIAL \
-            bazel run @graknlabs_build_tools//sonarcloud:code-analysis -- \
-            --organisation=$SONARCLOUD_ORGANISATION \
-            --project-key=$SONARCLOUD_PROJECT_KEY \
-            --commit-id=$CIRCLE_SHA1 \
-            --branch=$CIRCLE_BRANCH
+          echo "Skipping this job during development of this PR"
+          echo "TODO: Re-enable this job before submitting the PR"
+#          SONARCLOUD_CODE_ANALYSIS_CREDENTIAL=$SONARCLOUD_CREDENTIAL \
+#            bazel run @graknlabs_build_tools//sonarcloud:code-analysis -- \
+#            --organisation=$SONARCLOUD_ORGANISATION \
+#            --project-key=$SONARCLOUD_PROJECT_KEY \
+#            --commit-id=$CIRCLE_SHA1 \
+#            --branch=$CIRCLE_BRANCH
   
   test-common:
     machine: true
